restructured dir
