+++ src/pages/Work.vue
